Route mesh OpenAI traffic through unified openai proxy layer
