+++ mesh-llm/src/inference/election.rs
@@ -1183,6 +1183,7 @@ pub async fn election_loop(
     let mut currently_host = false;
     let mut current_local_port: Option<u16> = None;
     let mut llama_process: Option<launch::InferenceServerProcess> = None;
+    let mut backend_proxy: Option<crate::network::openai::backend::BackendProxyHandle> = None;
 
     // Initial settle
     tokio::time::sleep(std::time::Duration::from_secs(2)).await;
@@ -1380,6 +1381,9 @@ pub async fn election_loop(
                 } => {
                     eprintln!("🔄 [{}] llama-server died — restarting...", model_name);
                     llama_process = None;
+                    if let Some(proxy) = backend_proxy.take() {
+                        proxy.shutdown().await;
+                    }
                     currently_host = false;
                     current_local_port = None;
                     update_targets(&node, &model_name, InferenceTarget::None, &target_tx).await;
@@ -1400,6 +1404,9 @@ pub async fn election_loop(
         if currently_host {
             if let Some(process) = llama_process.take() {
                 process.handle.shutdown().await;
+            }
+            if let Some(proxy) = backend_proxy.take() {
+                proxy.shutdown().await;
             }
             tunnel_mgr.set_http_port(0);
             node.set_role(NodeRole::Worker).await;
@@ -1488,23 +1495,36 @@ pub async fn election_loop(
                 }
             };
 
+            let proxy = match crate::network::openai::backend::start_backend_proxy(llama_port).await
+            {
+                Ok(proxy) => proxy,
+                Err(err) => {
+                    eprintln!("  Failed to start local OpenAI backend proxy: {err}");
+                    process.handle.shutdown().await;
+                    on_change(true, false);
+                    last_worker_set = new_worker_set;
+                    let _ = peer_rx.changed().await;
+                    tokio::time::sleep(std::time::Duration::from_secs(3)).await;
+                    continue;
+                }
+            };
+            let local_proxy_port = proxy.port();
+            backend_proxy = Some(proxy);
+
             node.set_role(NodeRole::Host {
                 http_port: ingress_http_port,
             })
             .await;
-            // Point tunnel directly at llama-server, bypassing the API proxy.
-            // The client proxy has already normalized the request; the host
-            // doesn't need to re-parse or re-route it.
-            tunnel_mgr.set_http_port(llama_port);
+            tunnel_mgr.set_http_port(local_proxy_port);
             currently_host = true;
-            current_local_port = Some(llama_port);
+            current_local_port = Some(local_proxy_port);
             last_worker_set = new_worker_set;
             // Re-gossip so peers learn we're the host for this model
             node.regossip().await;
             update_targets(
                 &node,
                 &model_name,
-                InferenceTarget::Local(llama_port),
+                InferenceTarget::Local(local_proxy_port),
                 &target_tx,
             )
             .await;
@@ -1571,7 +1591,12 @@ pub async fn election_loop(
             } => {
                 eprintln!("🔄 [{}] llama-server died — restarting...", model_name);
                 llama_process = None;
+                if let Some(proxy) = backend_proxy.take() {
+                    proxy.shutdown().await;
+                }
                 currently_host = false;
+                current_local_port = None;
+                tunnel_mgr.set_http_port(0);
                 update_targets(&node, &model_name, InferenceTarget::None, &target_tx).await;
                 on_change(false, false);
             }
@@ -1590,6 +1615,9 @@ pub async fn election_loop(
     if currently_host {
         if let Some(process) = llama_process.take() {
             process.handle.shutdown().await;
+        }
+        if let Some(proxy) = backend_proxy.take() {
+            proxy.shutdown().await;
         }
         tunnel_mgr.set_http_port(0);
         node.set_role(NodeRole::Worker).await;
@@ -1627,6 +1655,7 @@ async fn moe_election_loop(
     let mut currently_running = false;
     let mut last_plan: Option<MoePlacementPlan> = None;
     let mut llama_process: Option<launch::InferenceServerProcess> = None;
+    let mut backend_proxy: Option<crate::network::openai::backend::BackendProxyHandle> = None;
     let mut current_local_port: Option<u16> = None;
     let mut last_plan_change_at = tokio::time::Instant::now();
 
@@ -1808,6 +1837,9 @@ async fn moe_election_loop(
             if let Some(process) = llama_process.take() {
                 process.handle.shutdown().await;
             }
+            if let Some(proxy) = backend_proxy.take() {
+                proxy.shutdown().await;
+            }
             tunnel_mgr.set_http_port(0);
             currently_running = false;
             current_local_port = None;
@@ -1875,13 +1907,32 @@ async fn moe_election_loop(
             .await
             {
                 Ok(process) => {
+                    let proxy = match crate::network::openai::backend::start_backend_proxy(
+                        llama_port,
+                    )
+                    .await
+                    {
+                        Ok(proxy) => proxy,
+                        Err(err) => {
+                            eprintln!("  Failed to start local OpenAI backend proxy: {err}");
+                            process.handle.shutdown().await;
+                            if peer_rx.changed().await.is_err() {
+                                break;
+                            }
+                            tokio::time::sleep(std::time::Duration::from_secs(3)).await;
+                            continue;
+                        }
+                    };
+                    let local_proxy_port = proxy.port();
+                    backend_proxy = Some(proxy);
+
                     node.set_role(NodeRole::Host {
                         http_port: ingress_http_port,
                     })
                     .await;
-                    tunnel_mgr.set_http_port(llama_port);
+                    tunnel_mgr.set_http_port(local_proxy_port);
                     currently_running = true;
-                    current_local_port = Some(llama_port);
+                    current_local_port = Some(local_proxy_port);
                     llama_process = Some(process);
                     if let Some(ref process) = llama_process {
                         on_process(Some(LocalProcessInfo {
@@ -1897,7 +1948,7 @@ async fn moe_election_loop(
                         &plan.fallback_ids,
                         my_id,
                         None,
-                        Some(llama_port),
+                        Some(local_proxy_port),
                         &model_name,
                     );
                     target_tx.send_replace(targets);
@@ -1958,13 +2009,29 @@ async fn moe_election_loop(
                 .await
                 {
                     Ok(process) => {
+                        let proxy =
+                            match crate::network::openai::backend::start_backend_proxy(llama_port)
+                                .await
+                            {
+                                Ok(proxy) => proxy,
+                                Err(err) => {
+                                    eprintln!(
+                                        "  Failed to start local OpenAI backend proxy: {err}"
+                                    );
+                                    process.handle.shutdown().await;
+                                    continue;
+                                }
+                            };
+                        let local_proxy_port = proxy.port();
+                        backend_proxy = Some(proxy);
+
                         node.set_role(NodeRole::Host {
                             http_port: ingress_http_port,
                         })
                         .await;
-                        tunnel_mgr.set_http_port(llama_port);
+                        tunnel_mgr.set_http_port(local_proxy_port);
                         currently_running = true;
-                        current_local_port = Some(llama_port);
+                        current_local_port = Some(local_proxy_port);
                         llama_process = Some(process);
                         if let Some(ref process) = llama_process {
                             on_process(Some(LocalProcessInfo {
@@ -1977,7 +2044,7 @@ async fn moe_election_loop(
                         update_targets(
                             &node,
                             &model_name,
-                            InferenceTarget::Local(llama_port),
+                            InferenceTarget::Local(local_proxy_port),
                             &target_tx,
                         )
                         .await;
@@ -2099,13 +2166,28 @@ async fn moe_election_loop(
             .await
             {
                 Ok(process) => {
+                    let proxy = match crate::network::openai::backend::start_backend_proxy(
+                        llama_port,
+                    )
+                    .await
+                    {
+                        Ok(proxy) => proxy,
+                        Err(err) => {
+                            eprintln!("  Failed to start local OpenAI backend proxy: {err}");
+                            process.handle.shutdown().await;
+                            continue;
+                        }
+                    };
+                    let local_proxy_port = proxy.port();
+                    backend_proxy = Some(proxy);
+
                     node.set_role(NodeRole::Host {
                         http_port: ingress_http_port,
                     })
                     .await;
-                    tunnel_mgr.set_http_port(llama_port);
+                    tunnel_mgr.set_http_port(local_proxy_port);
                     currently_running = true;
-                    current_local_port = Some(llama_port);
+                    current_local_port = Some(local_proxy_port);
                     llama_process = Some(process);
                     if let Some(ref process) = llama_process {
                         on_process(Some(LocalProcessInfo {
@@ -2121,7 +2203,7 @@ async fn moe_election_loop(
                         &plan.active_ids,
                         &plan.fallback_ids,
                         my_id,
-                        Some(llama_port),
+                        Some(local_proxy_port),
                         None,
                         &model_name,
                     );
@@ -2175,6 +2257,9 @@ async fn moe_election_loop(
     if currently_running {
         if let Some(process) = llama_process.take() {
             process.handle.shutdown().await;
+        }
+        if let Some(proxy) = backend_proxy.take() {
+            proxy.shutdown().await;
         }
         tunnel_mgr.set_http_port(0);
         node.set_role(NodeRole::Worker).await;
