+++ mesh-llm/src/network/mod.rs
@@ -1,5 +1,6 @@
 pub(crate) mod affinity;
 pub(crate) mod nostr;
+pub(crate) mod openai;
 pub(crate) mod proxy;
 pub(crate) mod rewrite;
 pub(crate) mod router;
