+++ mesh-llm/src/runtime/proxy/tests.rs
@@ -1,4 +1,6 @@
 use super::*;
+use crate::inference::pipeline;
+use crate::network::router;
 use crate::plugin;
 use crate::plugins::blobstore::BlobStore;
 use base64::Engine;
